services/voice/requirements.txt
# Voice Service Dependencies

# Core framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Environment configuration
python-dotenv==1.0.0

# For production Azure Speech integration (optional)
# azure-cognitiveservices-speech==1.31.0
